refactor(build): remove runtime-build-executor façade
Finalize Node outputs only in Rolldown, switch callers to installBuildRuntime, and align changelog/docs.
