+++ packages/core/package.json
@@ -196,10 +196,6 @@
 			"types": "./src/build/build-adapter.ts",
 			"default": "./src/build/build-adapter.ts"
 		},
-		"./build/runtime-build-executor": {
-			"types": "./src/build/runtime-build-executor.ts",
-			"default": "./src/build/runtime-build-executor.ts"
-		},
 		"./build/build-runtime": {
 			"types": "./src/build/build-runtime.ts",
 			"default": "./src/build/build-runtime.ts"
