+++ packages/core/src/services/module-loading/app-server-module-transpiler.service.test.ts
@@ -4,7 +4,7 @@ import { tmpdir } from 'node:os';
 import path from 'node:path';
 import { describe, it } from 'vitest';
 import { type BuildExecutor, setAppBuildAdapter, setAppBuildManifest } from '../../build/build-adapter.ts';
-import { installBuildRuntime } from '../../build/runtime-build-executor.ts';
+import { installBuildRuntime } from '../../build/build-runtime.ts';
 import { RolldownBuildAdapter } from '../../build/rolldown-build-adapter.ts';
 import type { EcoPagesElement } from '../../types/public-types.ts';
 import {
