+++ playground/kitchen-sink/bench/build-speed-bench.ts
@@ -1,6 +1,6 @@
 import path from 'node:path';
 import { bench, group } from 'mitata';
-import { installAppRuntimeBuildExecutor } from '../../../packages/core/src/build/runtime-build-executor';
+import { installBuildRuntime } from '../../../packages/core/src/build/build-runtime';
 import { createAppModuleLoader } from '../../../packages/core/src/services/module-loading/app-server-module-transpiler.service';
 import { resolveInternalExecutionDir } from '../../../packages/core/src/utils/resolve-work-dir';
 import { KITCHEN_SINK_REPRESENTATIVE_PAGES } from './lib/kitchen-sink-fixture';
@@ -12,7 +12,7 @@ const WARM_PAGES = Object.values(KITCHEN_SINK_REPRESENTATIVE_PAGES);
 
 async function importPageDiskWarm(filePath: string): Promise<void> {
 	const config = await loadStaticBuildBenchConfig();
-	installAppRuntimeBuildExecutor(config);
+	installBuildRuntime(config);
 	const loader = createAppModuleLoader(config);
 	const outdir = path.join(resolveInternalExecutionDir(config), '.server-modules');
